+++ src-new/sequelize.js
@@ -1,8 +1,24 @@
 import Case from 'case'
 
+export const files = ({ models, config = {} }) => ({
+  name: 'db',
+  files: [
+    { name: 'index.js', content: dbTemplate({ models, config }) },
+    {
+      name: 'models',
+      files: models.map(model => modelFile({ model, config }))
+    }
+  ]
+})
+
+const modelFile = ({ model, config }) => ({
+  name: `${Case.kebab(model.name)}.js`,
+  content: modelTemplate({ model, config })
+})
+
 // DB
 
-export const dbTemplate = ({ models = [], config = {} }) =>
+const dbTemplate = ({ models = [], config = {} }) =>
   `'use strict';
 
 const Sequelize = require('sequelize');
@@ -63,7 +79,7 @@ const renderNormalConnection = ({
 
 // Model
 
-export const modelTemplate = ({
+const modelTemplate = ({
   model: { name, fields = [], config = {} },
   config: {}
 }) =>
